+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.4.3
 * Fix sync axis by computing min and max category values
+* Move "sync axis" and "show only main axis" settings to "Axis" card
+* Add font setttings for measure units and axis
 
 ## 2.4.2
 * Check old enums with space like "Horizontal left" to maintain backward compatibility